Added new mkdocs configuration and updated project documentation
This commit introduces a new mkdocs configuration file for the multimodal-maestro project. This configuration includes details about the site, social links, navigation, theme, and plugins used.

In addition, a new index page for the documentation has been created, which provides a brief introduction to multimodal-maestro and installation instructions.

Changes to the pyproject.toml file include added dependencies required by the mkdocs configuration, as well as exclusions to the setuptools package find.

The README.md file is also updated, with more information about the project added, along with installation instructions